refactor(query): extract Canary.Query.Errors and collapse facade
Move from lib/canary/query.ex into lib/canary/query/errors.ex:
- Public: errors_by_service/3, errors_by_error_class/3, errors_by_class/1,
  error_detail/1, error_groups/1, error_summary/1
- Exposed as @doc false public for facade reuse: error_groups_since/1,
  error_summary_since/1
- Privates: apply_cursor/2, paginate_cursor/1, maybe_filter_annotation/2,
  select_group_with_classification/1, format_group/1,
  format_classification/1, classification_value/2, build_error_detail/1,
  group_summary/1, safe_decode_json/1 (all clauses)
- Module attr: @max_groups

Facade now contains only @moduledoc, report_slice/1, search/2, and
defdelegates. report_slice/1 calls Canary.Query.Errors.error_groups_since/1
and error_summary_since/1 directly. All Schemas aliases and
import Ecto.Query dropped from the facade.
